requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.1.2
djangorestframework==3.12.1
pycodestyle==2.6.0
pytz==2020.1
sqlparse==0.4.1
toml==0.10.1
